requirements.txt
keyboard==0.13.5
matplotlib==3.6.2
pafy==0.5.5
pip==23.1.2
python_vlc==3.0.18122
